Added some components
Created a directory-toolbar js,
classes tab js
classes toggle js,
and the css files to go with them.
Also to get it to work I ran this command [npm install react-bootstrap-validation --save] which eddited the json files, and i also added a folder 'node_modules'
I'm not sure what the command or the folder did, but i was having errors without them.